refactor: move libs to scripts/libs
Peer review modules are script-only; nesting under scripts/ lets
eslint-config-expensify/scripts rules apply without overrides.
